Move GenesisFiles module in Node namespace
