+++ lib/local-cluster/local-cluster.cabal
@@ -57,7 +57,7 @@ library
     Cardano.Wallet.Launch.Cluster.Faucet.Gen
     Cardano.Wallet.Launch.Cluster.Faucet.Serialize
     Cardano.Wallet.Launch.Cluster.FileOf
-    Cardano.Wallet.Launch.Cluster.GenesisFiles
+    Cardano.Wallet.Launch.Cluster.Node.GenesisFiles
     Cardano.Wallet.Launch.Cluster.Http.API
     Cardano.Wallet.Launch.Cluster.Http.Client
     Cardano.Wallet.Launch.Cluster.Http.Faucet.API
